Telemetry working Video working Controll transmission working All three with auto reconnect in case of connectivty drop in case of either drone or controller.
Introduced a shitty graphing solution that I just copy-pasted from github (just like 90% of this thing). This part needs work, it would be useful if able to plot altitude, speed and whatnot + plot the GPS possition eventually feeding back in telemetry over some sort of a map to aid in locating, avoiding obstacles while piloting since the camera feedback is limited to one point of view and locating in the event of a crash.